+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+##  [0.9.5]
+### Changed
+- The get_responses method to get_all_responses and changed it to return all responses
+- The CheckFirmwareVersion and SimpleAxis examples to use the new methods
+
+### Added
+- New get_response method to return a single response
+
+##  [0.9.1]
 ### Added
 - New API for the PTHat
 - version read only property.
@@ -17,5 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new setup files and uploaded project to PyPI
 
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.9.5...HEAD
+[0.9.5]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.9.1...v0.9.5
 [0.9.1]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.0...v0.9.1